# api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tech Book Store Practice API",
    "version": "1.0.0",
    "description": "A zero-dependency practice API for learning API test automation with Playwright. Seeded accounts: admin@bookstore.test / admin123 (admin) and user@bookstore.test / password123 (user). Call POST /api/reset any time to restore the original data.",
    "contact": {
      "name": "Tech Book Store Practice API"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local practice server"
    }
  ],
...
  "paths": {
...
    "/api/auth/apikey": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "API key protected endpoint",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Key accepted."
          },
          "401": {
            "description": "Missing or invalid API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/basic": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "HTTP Basic protected endpoint",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Basic auth accepted."
          },
          "401": {
            "description": "Missing or invalid Basic credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/limited": {
      "get": {
        "tags": [
          "Practice"
        ],
        "summary": "Rate-limited endpoint (429 after the limit)",
        "responses": {
          "200": {
            "description": "Within limit."
          },
          "429": {
            "description": "Rate limit exceeded. See Retry-After.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/delay": {
      "get": {
        "tags": [
          "Practice"
        ],
        "summary": "Delayed response (timeout practice)",
        "parameters": [
          {
            "name": "ms",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1000
            },
            "description": "Delay in milliseconds (max 10000)."
          }
        ],
        "responses": {
          "200": {
            "description": "Responded after the delay."
          }
        }
      }
    },
    "/api/status/{code}": {
      "get": {
        "tags": [
          "Practice"
        ],
        "summary": "Return a final HTTP status code from 200 to 599",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 200,
              "maximum": 599
            }
          }
        ],
        "responses": {
          "200": {
            "description": "For 2xx/3xx codes."
          },
          "400": {
            "description": "Code out of range.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/echo": {
      "get": {
        "tags": [
          "Practice"
        ],
        "summary": "Echo a GET request back",
        "responses": {
          "200": {
            "description": "Echo."
          }
        }
      },
      "post": {
        "tags": [
          "Practice"
        ],
        "summary": "Echo a POST request back",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Echo."
          }
        }
      },
      "put": {
        "tags": [
          "Practice"
        ],
        "summary": "Echo a PUT request back",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Echo."
          }
        }
      },
      "patch": {
        "tags": [
          "Practice"
        ],
        "summary": "Echo a PATCH request back",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Echo."
          }
        }
      },
      "delete": {
        "tags": [
          "Practice"
        ],
        "summary": "Echo a DELETE request back",
        "responses": {
          "200": {
            "description": "Echo."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Book 999 was not found."
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use the accessToken from login."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}